Chenges to .sh file, possible to give an argument 1 = delete build file and rebuild and do make otherwise run files from build file

+++ project/source_code/run.sh
@@ -1,11 +1,19 @@
 #!/usr/bin/env bash
 
-if [ -d "build" ]; then
-  rm -rf build
+A=$1
+
+if(( A == 1 )); then
+  if [ -d "build" ]; then
+    rm -rf build
+  fi
+  mkdir build
+  cd build
+  cmake ..
+  make
+else
+  cd build
 fi
-mkdir build
-cd build
-cmake ..
-make
+
+
 
 gnome-terminal --tab --title="input_handler" -e "./input_handler/input_handler" --tab --title="drivetrain" -e "./drivetrain/drivetrain" --tab --title="output_handler" -e "./output_handler/output_handler"
